splitting tweak file and adding a tweak
