+++ src/Cooked/Tweak.hs
@@ -3,15 +3,17 @@
 -- time using `Cooked.Ltl`
 module Cooked.Tweak (module X) where
 
-import Cooked.Tweak.AddInputsAndOutputs as X
 import Cooked.Tweak.Common as X hiding
   ( Tweak,
     UntypedTweak,
     runTweakInChain,
     runTweakInChain',
   )
+import Cooked.Tweak.Inputs as X
 import Cooked.Tweak.Labels as X
+import Cooked.Tweak.Mint as X
 import Cooked.Tweak.OutPermutations as X hiding (distinctPermutations)
+import Cooked.Tweak.Outputs as X
 import Cooked.Tweak.Signers as X
 import Cooked.Tweak.TamperDatum as X
 import Cooked.Tweak.ValidityRange as X
